+++ plat/arm/board/sgi575/platform.mk
@@ -9,6 +9,6 @@ include plat/arm/css/sgi/sgi-common.mk
 BL2_SOURCES		+=	lib/utils/mem_region.c			\
 				plat/arm/common/arm_nor_psci_mem_protect.c
 
-BL31_SOURCES		+=	lib/utils/mem_region.c			\
-				plat/arm/board/common/drivers/norflash/norflash.c \
+BL31_SOURCES		+=	drivers/cfi/v2m/v2m_flash.c		\
+				lib/utils/mem_region.c			\
 				plat/arm/common/arm_nor_psci_mem_protect.c
